+++ src/iosanita/policy/upgrades/v1002.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+
+from . import logger
+from plone import api
+from iosanita.policy.interfaces import IIoSanitaSettings
+
+import json
+
+
+def upgrade(setup_tool=None):
+    """ """
+    logger.info("Running upgrade (Python): Fix contatti_testata structure")
+    value = (
+        api.portal.get_registry_record("contatti_testata", interface=IIoSanitaSettings)
+        or ""
+    )
+    if not value:
+        return
+
+    value = json.loads(value)
+    value = [
+        {
+            "rootPath": "/",
+            "items": [value],
+        },
+    ]
+
+    api.portal.set_registry_record(
+        "contatti_testata", json.dumps(value), interface=IIoSanitaSettings
+    )